moved storage provider benchmarks to separate file
