+++ hub/core/tests/test_storage_provider.py
@@ -8,10 +8,6 @@ NUM_FILES = 20
 KEY = "file"
 
 
-mark_cache_group = pytest.mark.benchmark(group="storage_with_caches")
-mark_no_cache_group = pytest.mark.benchmark(group="storage_without_caches")
-
-
 # helper functions for tests
 def check_storage_provider(storage):
     FILE_1 = f"{KEY}_1"
@@ -110,60 +106,14 @@ def check_cache(cache):
     check_cache_state(cache, expected_state=[set(), set(), 0, 0, 0, 0])
 
 
-def write_to_files(storage):
-    chunk = b"0123456789123456" * MB
-    for i in range(NUM_FILES):
-        storage[f"{KEY}_{i}"] = chunk
-    storage.flush()
 
 
-def read_from_files(storage):
-    for i in range(NUM_FILES):
-        storage[f"{KEY}_{i}"]
-
-
-@mark_no_cache_group
 @parametrize_all_storages
 def test_storage_provider(storage):
     check_storage_provider(storage)
 
 
-@mark_cache_group
 @parametrize_all_caches
 def test_cache(storage):
     check_storage_provider(storage)
     check_cache(storage)
-
-
-@mark_no_cache_group
-@parametrize_all_storages
-def test_storage_write_speeds(benchmark, storage):
-    benchmark(write_to_files, storage)
-
-
-@mark_cache_group
-@parametrize_all_caches
-def test_cache_write_speeds(benchmark, storage):
-    benchmark(write_to_files, storage)
-
-
-@mark_no_cache_group
-@parametrize_all_storages
-def test_storage_read_speeds(benchmark, storage):
-    write_to_files(storage)
-    benchmark(read_from_files, storage)
-
-
-@mark_cache_group
-@parametrize_all_caches
-def test_cache_read_speeds(benchmark, storage):
-    write_to_files(storage)
-    benchmark(read_from_files, storage)
-
-
-@mark_cache_group
-@parametrize_all_caches
-def test_full_cache_read_speeds(benchmark, storage):
-    write_to_files(storage)
-    read_from_files(storage)
-    benchmark(read_from_files, storage)
